+++ lib/src/ilist/from_iterable_ilist_mixin.dart
@@ -1,3 +1,4 @@
+import "package:collection/collection.dart";
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 
 import "ilist.dart";